registration and login page with some change

+++ app/Http/Controllers/ThemeController.php
@@ -26,12 +26,23 @@ class ThemeController extends Controller{
 	}
 
 	public function dashboard(){
+		// if (Auth::check()) {
+		
 		$data['posts']=Post::limit(20)->offset(0)->latest()->get();
+		$data['authors']=Auth::user();
 
 		return view('dashboard', $data);
+		// }
+
+		// else{
+		// 	Session::flash('success_msg', 'Please login first to see dashboard');
+		// 	return view('auth/login');
+		// }
+
 	}
 
 	public function newpost(Request $request){
+		if (Auth::check()) {
 		if ($request->isMethod('post')) {
 			$post= new Post();
 			$post->title=$request->title;
@@ -47,7 +58,9 @@ class ThemeController extends Controller{
 			return redirect('dashboard');
 		}
 		return view('dashboard');
+			}
 	}
+
 	public function edit(Request $request, $id){
 		if ($request->isMethod('POST')) {
 			$edit=Post::find($id);
@@ -65,12 +78,22 @@ class ThemeController extends Controller{
 		// return "Hello"; die();
 		return view('edit', $data);
 	}
+	
 	public function delete($id){
-		Post::where('id', $id)->delete();
-
+		if(Auth::user()->userRole== 'Administrator'){
+			Post::where('id', $id)->delete();
 		Session::flash('success_msg', 'Your requested ID deleted successfully');
 		return redirect('dashboard');
+	}else{
+		Session::flash('success_msg', 'You have no permission to delete');
+		return redirect('dashboard');
+		}	
 	}
+
+	// public function authUpdate($id){
+	// 	return view('auth.auth_edit');
+	// }
+
 
 	public function gallery(){
 		$data['posts']=Post::paginate(5);
